+++ changelog.md
@@ -1,5 +1,17 @@
 # MODX Vapor Changelog
 
+## 1.0.0-beta-4 (October 8, 2012)
+
+- [#8817](http://tracker.modx.com/issues/8817) Fix media source import via resolvers
+- Fix invalid reference from $options to $fileMeta in resolvers
+- [#8871](http://tracker.modx.com/issues/8871) Fix import of extra tables without a table_prefix
+- Add excludeExtraTablePrefix to avoid prepending table_prefix for specific tables
+- Add excludeExtraTables option for excluding specific extra non-core tables
+- Add excludeFiles option for excluding specific files/dirs from base_path
+- Add ability to provide vaporOptions from config file
+
+## 1.0.0-beta-3 (September 26, 2012)
+
 - Fix PHP warning when table_prefix is empty
 - Fix realpath missing trailing slash on extension_packages